Move user tests to new package
